+++ src/px_misc.h
@@ -1,14 +1,14 @@
 extern void put_long_le(char *cp, long lval);
 extern long get_long_le(char *cp);
-extern int get_short_le(char *cp);
-extern void put_short_le(char *cp, int sval);
+extern short int get_short_le(char *cp);
+extern void put_short_le(char *cp, short int sval);
 extern void put_double_le(char *cp, double fval);
 extern double get_double_le(char *cp);
 
 extern void put_long_be(char *cp, long lval);
 extern long get_long_be(char *cp);
-extern int get_short_be(char *cp);
-extern void put_short_be(char *cp, int sval);
+extern short int get_short_be(char *cp);
+extern void put_short_be(char *cp, short int sval);
 extern void put_double_be(char *cp, double fval);
 extern double get_double_be(char *cp);
